Refactor let expressions to move optional vars
Elm strictly evaluates everything even if it's not used. So this moves
some "expensive" variables to places where they are guaranteed to be
used.

union got a pretty big refactor here: the setParent function was really
only there to try to make things more readible.  The jury is out on if I
like the way this turned out.