Add Computation module
A very simple monad implementation similar to Control.Monad.State. We'll
see if this turns out to be useful.

I was frustrated with the verbosity added by threading the state through
what would otherwise have been a simple `map` call when finding the
parent of a list of sets.

I'm not settled on the names of most of these functions yet.  I've used
the same names as the "original" haskell functions in a few cases but
I'd much rather use more idiomatically elm names (like `andThen`).